Add files via upload
Labels are used with MESEN-S

Original ISP patches lost, using patched ROMs instead.

Gun-A-Live modified every weapon in the game to be a gun, and every character is modified to be a gunslinger. This is more obvious in the endgame. However, checking item descriptions will show that every item that can be a weapon is now a gun according to the game.

"Every girl envies Sammo" was an early exercise in changing the addresses called with NPC dialogue. The result is every female NPC in one area of the game having the same line of dialogue (which originally belonged to a single NPC). Since NPC dialogue is called in several different places in the ROM (as well as RAM), this was an advancement in ROMhacking this game.